backend/requirements.txt
Django==4.2.27
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
djangorestframework-simplejwt==5.5.1
python-dotenv==1.0.0
Pillow==10.2.0
daphne==4.0.0
websockets==12.0
requests==2.31.0
openpyxl==3.1.2
setuptools>=65.0.0

